requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
mysql-connector-python==9.5.0
packaging==25.0
pillow==12.0.0
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
